pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "img2pdf>=0.6.1",
    "nest-asyncio>=1.6.0",
    "pdf2docx>=0.5.8",
    "pillow>=12.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "reportlab>=4.4.4",
    "yt-dlp>=2025.10.14",
]
